+++ packages/mobx/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* (Internal) Response handler receives `ResponseHandlerProps` object.
+
 ## 1.0.0-alpha.1
 
 * Switch from `action` to `navigation` (which contains an `action` as well as the previous `response`).